requirements.txt
# Existing dependencies
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
numpy==2.2.5
pandas==2.3.0
pydantic==2.11.5
pydantic-settings==2.2.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3

# New dependencies for production
fastapi-cache2==0.2.1
redis==5.0.1
aioredis==2.0.1
sentry-sdk==1.39.1
prometheus-client==0.19.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0